requirements.txt
python-telegram-bot[webhooks]>=20.0
python-dotenv==1.0.1
supabase>=1.0.0
requests==2.31.0
pytz==2024.1
uuid==1.30
pillow>=10.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
gunicorn>=20.1.0 